+++ CHANGELOG.md
@@ -8,6 +8,7 @@ PowerModels.jl Change Log
 - Update to InfrastructureModels v0.6 (#757) (breaking)
 - Update native power flow solvers to output results dict (#739) (breaking)
 - Ranamed `*_flow_cuts` to `_branch_power_cuts` (#713) (breaking)
+- Removed ref bus index from AdmittanceMatrix (#760) (breaking)
 - Dropped support for JuMP v0.20 (#673) (breaking)
 - Removed `_extend_pwl_cost!` function (#716)